requirements.txt
python-dotenv~=1.0.1
matplotlib~=3.9.0
pillow~=10.3.0
requests~=2.32.3
